requirements.txt
flask>=1.1.2
numpy>=1.19.5
pillow>=8.0.1
scikit-image>=0.17.2
torch>=1.7.0
torchvision>=0.8.1
waitress>=1.4.4
tqdm>=4.51.0
requests>=2.24.0
scipy>=1.5.4
pyheif>=0.5.1
pymatting>=1.1.1
filetype>=1.0.7
hsh>=1.1.0
